started organizing component folder
