+++ client/src/components/ProductDetail/productInformation.jsx
@@ -6,6 +6,9 @@ import Tracker from './imageGallery.jsx';
 import _ from 'underscore';
 import AddToCart from './addToCart.jsx';
 import EmptyStar from '../svgImages/EmptyStar.svg';
+import SizeAndQuantitySelector from './sizeAndQuantitySelector.jsx';
+import StyleSelector from './styleSelector.jsx';
+import CategoryName from './categoryName.jsx';
 
 // const productStyles = dummydata.productStyles;
 // const productInfo = dummydata.productInfo;
@@ -237,150 +240,5 @@ class ProductInformation extends React.Component {
   }
 }
 
-let CategoryName = function (props) {
-  return (
-    <Fragment>
-      {props.productInfo.map((productData) => {
-        return (
-          <div key={productData.id}>
-            <div className='product-category'>
-              {' '}
-              Category:{productData.category}
-            </div>
-            <h2 className='expanded-product-name'>{productData.name}</h2>
-          </div>
-        );
-      })}
-      {props.salesPrice === null ? (
-        <div className='product-category'>${props.originalPrice}</div>
-      ) : (
-        <div className='product-category'>
-          <span style={{ color: 'red' }}>${props.salesPrice}</span>{' '}
-          <span style={{ textDecoration: 'line-through' }}>
-            ${props.originalPrice}
-          </span>
-        </div>
-      )}
-    </Fragment>
-  );
-};
-
-let StyleSelector = function (props) {
-  let photos = props.photos;
-
-  let mappedStyles = props.sortedStyles.map((style) => {
-    return (
-      <div className='style-thumbnail-container' key={style.style_id}>
-        <FontAwesomeIcon
-          className={
-            props.checkedId === style.style_id
-              ? 'check-circle active-style'
-              : 'check-circle'
-          }
-          icon='check-circle'
-        />
-
-        <img
-          id={style.style_id}
-          name={style.name}
-          onClick={(e) => {
-            props.styleClickHandler(
-              e,
-              style.original_price,
-              style.sale_price,
-              style['default?']
-            );
-          }}
-          className={'style-thumbnail'}
-          src={style.photos[0].thumbnail_url}></img>
-      </div>
-    );
-  });
-
-  return (
-    <div className='style-header'>
-      Style{':'} <span className='style-name'>{props.defaultStyle}</span>
-      <div className='style-selector'>{mappedStyles}</div>
-    </div>
-  );
-};
-
-let SizeAndQuantitySelector = function (props) {
-  let currentSKU = _.pairs(props.SkusObj);
-
-  if (props.quantity > 0) {
-    var quantityRange = _.range(1, props.quantity + 1).filter(
-      (quantity) => quantity <= 15
-    );
-  } else {
-    quantityRange = [];
-  }
-
-  return (
-    <Fragment>
-      {props.sizeMenu > 1 ? (
-        <span className='invalid-cart-message'>Please select Size </span>
-      ) : null}
-      <div className='size-quantity-container'>
-        <select
-          // defaultValue='Select Size'
-          disabled={props.hasStock ? false : true}
-          className='size-selector'
-          size={props.sizeMenu}
-          onChange={(e) => props.sizeAndQuantityClickHandler(e)}>
-          {!props.hasStock ? (
-            <option className='size-default' id='disabled'>
-              OUT OF STOCK{' '}
-            </option>
-          ) : (
-            <option className='size-default'>{props.selectedSize}</option>
-          )}
-
-          {currentSKU
-            .filter((sku) => sku[1].size !== props.selectedSize)
-            .map((sku) => {
-              var styleObj = {
-                fontWeight: 'bold',
-                fontSize: 'medium',
-                display: 'flex',
-              };
-
-              if (sku[1].quantity === 0) {
-                styleObj.display = 'none';
-              }
-              return (
-                <option
-                  style={styleObj}
-                  key={sku[0]}
-                  id={sku[0]}
-                  data-quantity={sku[1].quantity}
-                  value={sku[1].size}>
-                  {sku[1].size}
-                </option>
-              );
-            })}
-        </select>
-
-        <select
-          onChange={props.quantityOnChange}
-          className='quantity-selector'
-          value={props.selectedQuantity}
-          disabled={!quantityRange.length ? true : false}>
-          {quantityRange.length === 0 ? (
-            <option>{'---'}</option>
-          ) : (
-            quantityRange.map((num) => {
-              return (
-                <option className='quantity-options' value={num} key={num}>
-                  {num}
-                </option>
-              );
-            })
-          )}
-        </select>
-      </div>
-    </Fragment>
-  );
-};
 
 export default ProductInformation;
